requirements.txt
seaborn
requests
tqdm
python-dotenv
halo
pandas
matplotlib
numpy